move to sdk/helper/testhelpers/schema
