+++ sdk/helper/testhelpers/schema/response_validation_test.go
@@ -1,4 +1,4 @@
-package testhelpers
+package schema
 
 import (
 	"testing"
